Move selection metrics calc (always minimized) to generation agg
